cmd/upload: document "raw" default for --format argument (#126)
When first using the tool, based on the `--help` output I did not
realize that `raw` was a supported format. Then, upon stumbling on a
GitHub issue that documents this format as being able to stream larger
images directly to disk, I found out that specifying `--format raw` does
not work, and leads to a failure relatively late in the image upload
process.

This documents that, when not specifying `--format`, a default format of
`raw` is assumed.